requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.2
cryptography==45.0.6
ecdsa==0.19.1
fastapi==0.116.1
grpcio==1.74.0
grpcio-reflection==1.74.0
grpcio-tools==1.74.0
h11==0.16.0
idna==3.10
jose==1.0.0
jwt==1.4.0
passlib==1.7.4
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
